requirements.txt
Django==4.0.2
virtualenv==20.13.1
django-crispy-forms==1.14.0
djangorestframework==3.13.1
mysqlclient==2.1.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
defusedxml==0.7.1
distlib==0.3.4
filelock==3.5.0
httpie==3.0.2
idna==3.3
multidict==6.0.2
Pillow==9.0.1
platformdirs==2.5.0
psycopg2==2.9.3
pygame==2.1.2
Pygments==2.11.2
PySocks==1.7.1
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
svgwrite==1.4.1
Tree==0.2.4
tzdata==2021.5
urllib3==1.26.8
